+++ app-specific-code/src/main/java/org/oskari/example/up/UPIndicatorsHandler.java
@@ -12,7 +12,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.oskari.example.IndicatorsUPHandler;
 import org.oskari.example.PostStatus;
 
 import fi.nls.oskari.annotation.OskariActionRoute;
@@ -24,6 +23,8 @@ import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
+import java.util.ArrayList;
+import org.oskari.example.UPTRoles;
 
 @OskariActionRoute("indicators_manager")
 public class UPIndicatorsHandler extends RestActionHandler {
@@ -34,7 +35,7 @@ public class UPIndicatorsHandler extends RestActionHandler {
 
     private static String upwsHost;
     private static String upwsPort;
-    private static final Logger log = LogFactory.getLogger(IndicatorsUPHandler.class);
+    private static final Logger log = LogFactory.getLogger(UPIndicatorsHandler.class);
     
     private JSONArray errors;
     private ObjectMapper Obj;
@@ -66,6 +67,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "SELECT id, indicator\n"
                     + "	FROM public.up_indicators;"
@@ -97,6 +103,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "INSERT INTO public.up_indicators(\n"
                     + "	indicator)\n"
@@ -125,6 +136,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "UPDATE public.up_indicators\n"
                     + "	SET indicator=?\n"
@@ -154,6 +170,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM public.up_indicators\n"
                     + "	WHERE id=?;"
